Issue #16: Modified ProfInfo
The _score instance variable is actually not needed, since we are going to be
calculating the score using the functors. What we actually need are the total
time the professor spends on campus, and the total number of days the
professor spends on campus. I included code that calculates these variables
in the addTime() method.