Issue #16: ProfInfo stores more detailed info
Prior to this change, ProfInfo had to require that times were being added to the schedule sequentially, by Weekday/TimeBlock. I have refactored the class to remove this requirement. There is now a vector<char> that holds 0 if the professor comes to class on a Weekday (char type because vector<bool> is "broken"), and a vector< pair<TimeBlock, TimeBlock> > that holds the first and last TimeBlocks that the Prof spends on campus.