ScheduleGenerator.h fixes
ScheduleGenerator is an abstract base class. It is never meant to be
instantiated, so its constructor should not be made public. The member
vectors are meant to be used by subclasses; if all subclasses don't use
them, we should probably remove them from the base class altogether.

In any case, this is a design issue that requires group input.