Issue #9: Created base class for testing.
This is an abstract base class. Subclasses will be created for every concrete class, and also for main(). Each subclass will call the test methods for every other concrete class that it will deal with; for example, the tester for main() will create and call test objects for the FileReaderFactory; the test for FileReaderFactory will call the tests for each of the concrete FileReader objects; and so forth, down the component heirarchy.