Issue #9: Created all needed test classes
This should be all of the test classes we need. All concrete classes in
the project now have a related test class.

The test classes, at the moment, have only the default implementations
for testPass() and testFail(), inherited from the base TestRunner class,
and do nothing except print that tests are not written. They may be
implemented as each of the classes that need to be tested (like
GreedyScheduleGenerator) can compile.