Issue #9: Fixed comments
There's a "trick" you can use to comment out code. At the end of every
block you may or may not comment out, put this:
//*/

Then, at the start of the block, put this:
//*

The block will not be commented, and when you do want to comment it, you
just delete the first slash symbol.

Unfortunately, you have to be careful not to accidentally nest these
things, or GCC complains. I was not so careful.