Issue #9: Created DataCreator.loadProfs()
That way, the vector of Course objects will have professor ID's that are
guaranteed to match a specific vector of Prof objects. Not that this
still doesn't guarantee that the Course vector is valid; it does not
check how many courses a professor teaches, and so may assign more than
three to one professor.