Issue #9: DataCreator can create a Schedule
The Schedule that is created is valid insofar as room capacity is
concerned. It will not have been checked to make sure that the other
constraints are valid (e.g. that courses taught by the same Prof are not
taught at the same time). The way that the Schedule is generated makes
this unlikely, but possible.