Issues #9 and #72: Wrote Schedule tests
To figure out what was causing Schedule to crash, I wrote more rigorous
test for the class. Once I found out the reason, I changed the pass test
to avoid it. It is still in the fail test.

I found out that DataCreator::createSchedule() can also cause this
error, so I fixed it. (It is literally a one-character fix.)