Moved the ProfInfo map to method
Instead of having to copy, pass, and delete different ProfInfo maps -
and all the confusion that entails - I realized that it would be much
smoother if I just moved it into calculateScore(). (This was how I had
it originally.) The only disadvantage is that it would not persist
across method calls, meaning that we can't generate the score for only
one professor at a time. However, it seems like it is just as efficient
to do the whole schedule at once, so this is not really an issue.

The method itself works as expected. It is still crashing when we delete
things after swapping (and only after swapping), but the only thing it's
deleting now is the Schedule, so I know that it's a problem with that
class.