Issue #9: Unit testing possible
The main() function now accepts a command-line argument to make unit
testing possible. The switch is "-c" followed by the name of the
(concrete) class you want to test.

For example, if you want to only test the GeneticScheduleGenerator, and
not any of the subtests, you would pass "-c GeneticScheduleGenerator" on
the command line.