Fixed #89
When I was accounting for zero indexing, I was considering "pos" to be
the position of the last character. But this is not the `substr`
function takes; it takes a length. Forgot to set it back, resulting in
edge cases where the ellipsed string is one character too long. Fixed.