Change of "-t" opinion
I feel it may be not good to let user to wait about 1 or more minutes
for generating schedule. So I change the unit of -t option to let user
able to set the time limitation in seconds, not minutes.  If they don't
want to wait 1 min, they can set to 20 or 30 seconds.